refactor(all): 폴더 구조 변경
- 테스트 폴더 구조를 src 내 구조와 동일하게 변경
- utils 하위 폴더 제거
- CustomError를 utils로 이동
