+++ __tests__/domains/LottoPurchaseManagerTest.js
@@ -1,5 +1,5 @@
-import LottoPurchaseManager from '../src/domains/LottoPurchaseManager.js';
-import Lotto from '../src/domains/Lotto.js';
+import LottoPurchaseManager from '../../src/domains/LottoPurchaseManager.js';
+import Lotto from '../../src/domains/Lotto.js';
 
 describe('로또 구매 테스트', () => {
   test.each([
